# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Películas y Categorías",
        "description": "Documentación de la API para gestionar películas y categorías",
        "version": "1.0"
    },
    "paths": {
        "/api/categorias": {
            "get": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Lista de categorías",
                "description": "Obtener todas las categorías.",
                "operationId": "66bfb3e0bf7747700be142ec5329dfae",
                "responses": {
                    "200": {
                        "description": "Lista de categorías obtenida con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Categoria"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Agregar una categoría",
                "description": "Almacenar una nueva categoría.",
                "operationId": "acace4e67a5f9bbaa41b456c95b2bd05",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Nombre de la categoría",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Categoría agregada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categoria"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/categorias/{id}": {
            "get": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Obtener una categoría especifica",
                "description": "Obtener una Categoria por ID.",
                "operationId": "af067e125b5f9a1b1306d9e5c2a5867e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoría obtenida con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categoria"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Buscar y actualizar una categoría",
                "description": "Actualizar una categoría existente.",
                "operationId": "3465ccc7d2520d34b4c457d4158ae6ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Nueva Categoría"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Datos actuales de la categoría antes de editar y actualización exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categoria"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Eliminar la información de una categoría existente.",
                "operationId": "a37893cb60d6b5ca9b1d6d23dbe83041",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "NO CONTENT"
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se pudo realizar correctamente la operación"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Categorías"
                ],
                "summary": "Actualizar parcialmente una categoría",
                "description": "Actualizar parcialmente una categoría existente.",
                "operationId": "65e7d7c83ce9622baf30f0102f1e0b6d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Nuevo Nombre"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categoría actualizada parcialmente con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Categoria"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Categoria": {
                "title": "Categoría",
                "description": "Modelo de una categoría",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Acción"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}